use fake slack
