+++ tests/apps/github_stats/conftest.py
@@ -12,25 +12,3 @@ def create_user(session):
         return new_user
 
     return _new_user
-
-
-@pytest.fixture
-def fake_slack_client(mocker):
-    class FakeSlackClient:
-        def __init__(self, *, channel_info=None):
-            self.mock = mocker.MagicMock()
-            if channel_info:
-                self.channel_info = channel_info
-
-        def get_channel_info(self, *args, **kwargs):
-            self.mock(*args, **kwargs)
-            return self.channel_info
-
-        def post_message(self, *args, **kwargs):
-            self.mock(*args, **kwargs)
-            return
-
-        def __repr__(self):
-            return "<FakeSlackClient>"
-
-    return FakeSlackClient
